+++ Cems-Client/src/clientControllers/LoginScreenController.java
@@ -8,9 +8,6 @@ import clientHandlers.ClientUI;
 import clientHandlers.ClientHandler;
 import javafx.event.ActionEvent;
 import javafx.fxml.FXML;
-import javafx.fxml.FXMLLoader;
-import javafx.scene.Parent;
-import javafx.scene.Scene;
 import javafx.scene.control.Button;
 import javafx.scene.control.Label;
 import javafx.scene.control.PasswordField;
