+++ Cems-Client/src/clientFXMLS/HOD.fxml
@@ -8,7 +8,7 @@
 <?import javafx.scene.layout.VBox?>
 <?import javafx.scene.text.Font?>
 
-<AnchorPane maxHeight="-1.0" maxWidth="-1.0" prefHeight="720.0" prefWidth="1280.0" stylesheets="@../gui/Stylesheet.css" xmlns="http://javafx.com/javafx/19" xmlns:fx="http://javafx.com/fxml/1" fx:controller="clientControllers.HODController">
+<AnchorPane maxHeight="-1.0" maxWidth="-1.0" prefHeight="720.0" prefWidth="1280.0" stylesheets="/gui/Stylesheet.css" xmlns="http://javafx.com/javafx/19" xmlns:fx="http://javafx.com/fxml/1" fx:controller="clientControllers.HODController">
 <children>
   <Label alignment="CENTER" layoutX="150.0" layoutY="178.0" prefHeight="36.0" prefWidth="332.0" style="&#10;" text="Welcome back , User !" textAlignment="CENTER" wrapText="false" AnchorPane.leftAnchor="150.0" AnchorPane.rightAnchor="150.0">
     <font>
