+++ Cems-Client/src/logic/QuestionModel.java
@@ -1,17 +1,8 @@
 package logic;
 
-import java.io.IOException;
 import java.io.Serializable;
 
-import clientControllers.EditQuestionController;
-import javafx.event.ActionEvent;
-import javafx.event.EventHandler;
-import javafx.fxml.FXMLLoader;
-import javafx.scene.Node;
-import javafx.scene.Scene;
 import javafx.scene.control.Button;
-import javafx.scene.layout.AnchorPane;
-import javafx.stage.Stage;
 
 public class QuestionModel implements Serializable {
 
@@ -27,41 +18,6 @@ public class QuestionModel implements Serializable {
         this.questiontext = questiontext;
         this.questionnumber = questionnumber;
         this.lecturer = lecturer;
-    }
-
-    public void createButton() {
-        edit = new Button("Edit");
-
-        edit.setOnAction(new EventHandler<ActionEvent>() {
-
-            @Override
-            public void handle(ActionEvent event) {
-                // remember all the data of the question
-                // send to EditQuestionScreen
-                ((Node) event.getSource()).getScene().getWindow().hide();
-                AnchorPane root = null;
-                Stage currentStage = new Stage();
-                FXMLLoader loader = new FXMLLoader();
-                try {
-                    root = loader.load(getClass().getResource("/clientFXMLS/EditQuestion.fxml").openStream());
-                } catch (IOException e) {
-                    e.printStackTrace();
-                }
-                
-                EditQuestionController eqc = loader.getController();
-                eqc.loadQuestion(getQuestion(), getId());
-                System.out.println("opening edit question");
-                Scene scene = new Scene(root);
-                scene.getStylesheets().add(getClass().getResource("/clientFXMLS/background.css").toExternalForm());
-                currentStage.setScene(scene);
-                currentStage.setTitle("CEMS System - Lecturer - Edit Question");
-                currentStage.show();
-            }
-        });
-
-        edit.setId("editbutton");
-        edit.setPrefWidth(60);
-        edit.setPrefHeight(20);
     }
 
     public String getId() {
@@ -113,8 +69,7 @@ public class QuestionModel implements Serializable {
     }
 
     public Button getEdit() {
-        createButton();
-        return this.edit;
+        return edit;
     }
 
     public void setEdit(Button edit) {
