+++ Cems-Server/src/serverControllers/ServerController.java
@@ -60,7 +60,7 @@ public class ServerController {
         hostNameColumn.setCellValueFactory(new PropertyValueFactory<>("Host"));
         ipColumn.setCellValueFactory(new PropertyValueFactory<>("Ip"));
 
-        passwordField.setText("1234");
+        passwordField.setText("02587595mM!");
         usernameField.setText("root");
         dbNameField.setText("projecton");
         try {
